Make docs more consistent
In particular, use "web service" rather than "endpoint" everywhere. This
more closely matches our other documentation. Also, removed comments
about a value being available from all databases and web services as
we did so inconsistently and it was noise. We now only mention if it
is limited to particular databases or web services.